Supporting additional loaders
CSS loaders. Looks like webpack removed support for addDepedency so if
you edit a css file that is imported, the app won't hot reload unless
you also make an edit to the main css file that is referenced. Hopefully
the import loaders we use will work with whatever changes have been made
in webpack 2 to support this.